Refactor controllers website_school_portal
