+++ website_school_portal/controllers/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from . import controllers
+from . import snippets
 from . import reservations
+from . import mes_donnees
